Improved rendering speed
The rendering speed is improved for images with part of the image off screen by moving the valid pixel check out of xy_set and into every individual function that sets pixels. The one that displays "jayde the goo deer" has been improved by 32% because it renders so much off screen all the time.